pyproject.toml
[project]
name = "analyzing-baseball-data-with-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ibis-framework>=9.5.0",
    "jupyterlab>=4.3.3",
    "ruff>=0.8.3",
]
